% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_isotope_pairs.R
\name{create_isotope_pairs}
\alias{create_isotope_pairs}
\title{Create isotope pairs}
\usage{
create_isotope_pairs(isotope_n = NULL, isotope_names = NULL)
}
\arguments{
\item{isotope_n}{number of isotopes}

\item{isotope_names}{names of isotopes}
}
\description{
Create isotope pairs
}
