% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-extract_overlap.R
\name{extract_overlap}
\alias{extract_overlap}
\title{extract overlap}
\usage{
extract_overlap(data, name_a = NULL, name_b = NULL)
}
\arguments{
\item{data}{a \code{array} object containing \code{matrices} created by the function
\code{overlap()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.}

\item{name_a}{character string to supply for the first
\code{sample_name} used in \code{overlap()}. Defaults to \code{"sample_name_a"}.}

\item{name_b}{character string to supply for the second
\code{sample_name} used in \code{overlap()}. Defaults to \code{"sample_name_b"}.}
}
\value{
A \code{tibble} containing five rows, \code{sample_name_a}, \code{id},
\code{sample_name_b}, \code{sample_number}, and \code{niche_overlap}.
}
\description{
Extract Bayesian estimates of similarities among groups produced by the
following function \code{overlap()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
extract_overlap(data = over_stat)

}
\seealso{
\code{\link[nicheROVER:overlap]{nicheROVER::overlap()}}
}
