% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-extract_sigma.R
\name{extract_sigma}
\alias{extract_sigma}
\title{extract Σ}
\usage{
extract_sigma(data, isotope_a = NULL, isotope_b = NULL, format = "wide")
}
\arguments{
\item{data}{a \code{list} created by the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}}

\item{isotope_a}{\code{character} string to supply for the first
isotope used in \code{niw.post()}. Defaults to \code{"d15n"}.}

\item{isotope_b}{\code{character}string to supply for the second
isotope used in \code{niw.post()}. Defaults to \code{"d13c"}.}

\item{format}{format a \code{character} that decided whether the returned object is
in long or wide format. Default is \code{"wide"}, with the alternative supplied
being \code{"long"}.}
}
\value{
Returns a \code{tibble} of extracted estimates of Σ created by the
function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
The tibble will contain five columns in the following order, \code{metric}, \code{id},
\code{sample_name}, \code{isotope}, \code{sample_number}, and the posterior sample for Σ
(i.e., \code{post_sample}).
}
\description{
Extract Bayesian estimates of Σ from the function \code{niw.post()} in the
package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
extract_sigma(
data = niw_fish_post
)

}
\seealso{
\code{\link[nicheROVER:niw.post]{nicheROVER::niw.post()}}
}
