% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu_est_long}
\alias{mu_est_long}
\title{Estimates of μ from Normal-Inverse-Wishart (NIW) prior}
\format{
data frame containing 8,000 rows and 7 variables
\describe{
\item{metric}{name of the metric extracted from \code{niw.post()}}
\item{species}{species abbreviation}
\item{sample_number}{sample number from 1-1000}
\item{isotope}{column with isotope name}
\item{mu_est}{estimate of mu produced from \code{niw.post()}}
\item{element}{isotopic element used in labelling}
\item{neutron}{neutron number used in labelling}
}
}
\usage{
mu_est_long
}
\description{
Estimates of μ from Normal-Inverse-Wishart (NIW) priors for example data
frame from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\keyword{datasets}
