% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mu_extract.R
\name{mu_extract}
\alias{mu_extract}
\title{Mu extract}
\usage{
mu_extract(data)
}
\arguments{
\item{data}{a list object created by \code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}}
}
\description{
Extract Bayesian estimates of mue from the function \code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
df_mu <- mu_extract(
data = niw_fish_post
)

}
