% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-niche_size_extract.R
\name{niche_size_extract}
\alias{niche_size_extract}
\title{Niche Size Extract}
\usage{
niche_size_extract(data, name = NULL, prob = NULL)
}
\arguments{
\item{data}{a \code{list} created by the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.}

\item{name}{a \code{character} string that will be assinged as the column name for
groups. Default is \code{sample_name}.}

\item{prob}{a \code{numeric} bound by 0 and 1 indicating the
probabilistic niche size. Default is 0.95.}
}
\value{
a \code{tibble} containing three rows, \code{sample_name}, \code{id}, and \code{niche_size}.
}
\description{
Extract niche size based on elliptical niche region of Bayesian estimates of
sigma created by function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
This function is a wrapper around \code{nicheROVER::niche.size}.
}
\examples{
niche_size <- niche_size_extract(data = niw_fish_post)


}
\seealso{
\code{\link[nicheROVER:niche.size]{nicheROVER::niche.size()}} \code{\link[nicheROVER:niw.post]{nicheROVER::niw.post()}}
}
