% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{niw_fish_post}
\alias{niw_fish_post}
\title{Estimates of mu and sigma from Normal-Inverse-Wishart (NIW) prior}
\format{
A list with elements mu and Sigma of sizes c(nsamples, length(lambda))
and c(dim(Psi).
}
\usage{
niw_fish_post
}
\description{
Estimates of sigma from Normal-Inverse-Wishart (NIW) priors for example data
frame produced using \code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}
}
\keyword{datasets}
