% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{post_sam_siber}
\alias{post_sam_siber}
\title{A \code{list} of the posterior estimates of \eqn{\mu} and \eqn{\Sigma} from \code{{SIBER}}}
\format{
A \code{list} with estimates of \eqn{\mu} and \eqn{\Sigma} for each species
and group.
}
\usage{
post_sam_siber
}
\description{
Posterior estimates of \eqn{\mu} and \eqn{\Sigma} using the \code{demo.siber.data}
data set from \href{https://cran.r-project.org/web/packages/SIBER/index.html}{{SIBER}}.
This \code{list} is produced  using the function \code{siberMVN()} from
\href{https://cran.r-project.org/web/packages/SIBER/index.html}{{SIBER}}.
}
\keyword{datasets}
