% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_ellipse.R
\name{sigma_ellipse}
\alias{sigma_ellipse}
\title{Create ellipses based on Bayesian estimates}
\usage{
sigma_ellipse(
  dat_mu,
  dat_sigma,
  name = "sample_name",
  number = "sample_number",
  mu_name = "mu_est",
  isotope_a = "cal_d15n",
  isotope_b = "cal_d13c",
  p_ell = NULL
)
}
\arguments{
\item{dat_mu}{a data frame containing mu Bayesian estimates. This data frame
needs to be in long format with each mu estimate for each isotope stacked
on top of each each other.}

\item{dat_sigma}{a data frame containing sigma Bayesian estimates.
This data frame needs be in wide format that is sigma/covariance matrices stacked
onto of each other.}

\item{name}{character string that is the column name of each grouping used
(e.g, species) to create mu and sigma Bayesian estimates. This name is the same for
\code{dat_mu}and \code{dat_sigma}. Defaults to \code{"sample_name"}.}

\item{number}{character string that is the column name of each Bayesian estimate.
This name is the same for \code{dat_mu}and \code{dat_sigma}. Defaults to \code{"sample_number"}.}

\item{mu_name}{character string that is the column name of the mu estimate in
\code{dat_mu}. Defaults to \code{"mu_est"}.}

\item{isotope_a}{character string that is the column name of the first isotope.
This name is the same for \code{dat_mu}and \code{dat_sigma}. Defaults to \code{"cal_d15n"}.}

\item{isotope_b}{character string that is the column name of the first isotope.
This name is the same for \code{dat_mu}and \code{dat_sigma}. Defaults to \code{"cal_d13c"}.}

\item{p_ell}{is the confidence interval of each ellipse estimate.
Default is 0.95 (i.e., 95\% confidence interval).}
}
\value{
a data frame containing 100 estimates for each \code{sample_number} of each
\code{sample_name}. As the function runs, it will produce progress bars and tell
the user what step it is on. After the last step it will complete a few more
processes and then return the total time the function ran.
}
\description{
This function allows the user to supply Bayesian estimates of mu and sigma to
create estimated Bayesian ellipse for niche region.
}
\examples{
# ellipse <- sigma_ellipse()

}
