% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-sigma_extract.R
\name{sigma_extract}
\alias{sigma_extract}
\title{Sigma extract}
\usage{
sigma_extract(data, isotope_a = NULL, isotope_b = NULL)
}
\description{
Extract estiamtes of sigma from functions from {nicheROVER}.
}
