% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layman.R
\name{extract_layman}
\alias{extract_layman}
\title{extract Layman metrics}
\usage{
extract_layman(
  data,
  community_df = NULL,
  community_name = NULL,
  data_format = NULL
)
}
\arguments{
\item{data}{siber object}

\item{community_name}{a character string that is the name of the second column
in community_df supplied (e.g., \code{"region"}).}

\item{coummunity_df}{a two column dataframe one of the column names
has to be a number as a character which is the same community names requried
by \code{makeSiberObject()} function from {SIBER}. The other column is the actual
names of the communities that the user is working with (e.g., \code{"region"})}
}
\description{
Extract Bayesian estimates of for the following six layman metrics,
\eqn{\delta^{13}}C from data objects created by
\href{https://cran.r-project.org/web/packages/SIBER/index.html}{{SIBER}}. These six
metric include the following Layman  of dX_range, dY_range, TA,
CD, MNND and SDNND
}
