% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-extract_mu.R
\name{extract_mu}
\alias{extract_mu}
\title{extract μ}
\usage{
extract_mu(data, format = "long")
}
\arguments{
\item{data}{a \code{list} created by the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.}

\item{format}{a \code{character} that decided whether the returned object is
in long or wide format. Default is \code{"long"}, with the alternative supplied
being \code{"wide"}.}
}
\value{
Returns a \code{tibble} of extracted estimates of μ that are created by
the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
The \code{tibble} will contain five columns in the following order, \code{metric},
\code{sample_name}, \code{sample_number}, and the names of the isotope
columns supplied to \code{niw.post()} (e.g., \code{d15n} and \code{d13c}).
}
\description{
Extract Bayesian estimates of μ from the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
extract_mu(
data = niw_fish_post
)

}
\seealso{
\code{\link[nicheROVER:niw.post]{nicheROVER::niw.post()}}
}
