% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu_est_long}
\alias{mu_est_long}
\title{A \code{data.frame} containing posterior estimates of μ}
\format{
\code{data.frame} containing 8,000 rows and 7 variables
\describe{
\item{metric}{name of the metric extracted from \code{niw.post()}}
\item{species}{species abbreviation}
\item{sample_number}{sample number from 1-1000}
\item{isotope}{column with isotope name}
\item{mu_est}{estimate of mu produced from \code{niw.post()}}
\item{element}{isotopic element used in labelling}
\item{neutron}{neutron number used in labelling}
}
}
\usage{
mu_est_long
}
\description{
Posterior estimates of μ using \code{fish} data set from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}},
using  Normal-Inverse-Wishart (NIW) priors.
}
\keyword{datasets}
