% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mu_extract.R
\name{mu_extract}
\alias{mu_extract}
\title{μ extract}
\usage{
mu_extract(data)
}
\arguments{
\item{data}{a \code{list} created by the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.}
}
\value{
Returns a \code{tibble} of extracted estimates of μ that are created by
the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
The \verb{tibble will} contain five columns in the following order, \code{metric},
\code{sample_name}, \code{sample_number}, and the names of the isotope
columns supplied to \code{niw.post()} (e.g., \code{d15n} and \code{d13c}).
}
\description{
Extract Bayesian estimates of μ from the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
mu_extract(
data = niw_fish_post
)

# ---- below will be turned into vignette when finished ----
\dontrun{
# load packages
{
  library(dplyr)
  library(janitor)
  library(nicheROVER)
  library(nichetools)
  library(purrr)
 }

# get fish data frame, and remove sulfur for the example
df <- fish \%>\%
    janitor::clean_names()

# create number of samples used in nicheROVER
nsample <- 1000

# split the data frame by species and select isotopes of interest
df_split <- df \%>\%
  split(.$species) \%>\%
  map(~ select(., d15n, d13c))

# extract the names of each list to name each object in list
df_names <- df \%>\%
   group_by(species) \%>\%
   group_keys() \%>\%
   ungroup() \%>\%
   mutate(
     id = 1:nrow(.)
     )

# name each object in list
names(df_split) <- df_names$species

# create niw posterior samples
niw_fish_post <- df_split \%>\%
  map(~niw.post(nsample = nsample, X = .))

# ---- extract mu estimates ----
mu_extract(
data = niw_fish_post
)
}

}
\seealso{
\code{\link[nicheROVER:niw.post]{nicheROVER::niw.post()}}
}
