% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-niche_ellipse.R
\name{niche_ellipse}
\alias{niche_ellipse}
\title{Create ellipses based on Bayesian estimates}
\usage{
niche_ellipse(
  dat_mu,
  dat_sigma,
  isotope_a = NULL,
  isotope_b = NULL,
  p_ell = NULL
)
}
\arguments{
\item{dat_mu}{a data frame containing mu Bayesian estimates. This data frame
needs to be in long format with each mu estimate for each isotope stacked
on top of each each other.}

\item{dat_sigma}{a data frame containing sigma Bayesian estimates.
This data frame needs be in wide format that is sigma/covariance matrices stacked
onto of each other.}

\item{isotope_a}{character string that is the column name of the first
isotope used in \code{dat_sigma}. Defaults to \code{"d15n"}.}

\item{isotope_b}{character string that is the column name of the second
isotope used in \code{dat_sigma}. Defaults to \verb{d13c"}.}

\item{p_ell}{is the confidence interval of each ellipse estimate.
Default is 0.95 (i.e., 95\% confidence interval).}
}
\description{
This function allows the user to supply Bayesian estimates of mu and sigma to
create estimated Bayesian ellipse for niche region.
}
\examples{
# ellipse <- sigma_ellipse()

}
