% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichetools-package.R
\docType{package}
\name{nichetools-package}
\alias{nichetools}
\alias{nichetools-package}
\title{nichetools: Complementary package to `{nicheROVER}`}
\description{
This package is complementary to {nicheROVER} as it allows the user to extract Bayesian estimates from data objects created by package {nicheROVER}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://benjaminhlina.github.io/nichetool/}
}

}
\keyword{internal}
