% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichetools-package.R
\docType{package}
\name{nichetools-package}
\alias{nichetools}
\alias{nichetools-package}
\title{nichetools: Complementary Package to 'nicheROVER' and 'SIBER'}
\description{
Provides functions complementary to packages 'nicheROVER' and 'SIBER' allowing the user to extract Bayesian estimates from data objects created by the packages 'nicheROVER' and 'SIBER'. Please see the following publications for detailed methods on 'nicheROVER' and 'SIBER' Hansen et al. (2015) \doi{10.1890/14-0235.1}, Jackson et al. (2011) \doi{10.1111/j.1365-2656.2011.01806.x}, and Layman et al. (2007) \doi{10.1890/0012-9658(2007)88[42:CSIRPF]2.0.CO;2}, respectfully.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://benjaminhlina.github.io/nichetools/}
  \item \url{https://github.com/benjaminhlina/nichetools}
  \item Report bugs at \url{https://github.com/benjaminhlina/nichetools/issues}
}

}
\author{
\strong{Maintainer}: Benjamin L. Hlina \email{benjamin.hlina@gmail.com}

}
\keyword{internal}
