% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichetools-package.R
\docType{package}
\name{nichetools-package}
\alias{nichetools}
\alias{nichetools-package}
\title{nichetools: Complementary Package to NicheROVER and SIBER}
\description{
Provides functions complementary to packages nicheROVER and SIBER allowing the user to extract Bayesian estimates from data objects created by packages nicheROVER and SIBER.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://benjaminhlina.github.io/nichetools/}
}

}
\keyword{internal}
