% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{niw_fish_post}
\alias{niw_fish_post}
\title{A \code{list} of the posterior estimates of \eqn{\mu} and \eqn{\Sigma} from {nicheROVER}}
\format{
A \code{list} with elements \eqn{\mu} and \eqn{\Sigma} of sizes c(nsamples, length(lambda))
and c(dim(Psi)).
}
\usage{
niw_fish_post
}
\description{
Posterior estimates of \eqn{\mu} and \eqn{\Sigma} using the \code{fish} data set from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}},
using Normal-Inverse-Wishart (NIW) priors.
This \code{list} is produced  using the function \code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\keyword{datasets}
