% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{over_stat}
\alias{over_stat}
\title{A \code{data.frame} of estimates of percentage of overlap among groups produced by
\code{niw.post()}}
\format{
A \code{array}containing \code{matrices} of the percent overlap for each group
used in Bayesian estimates of μ and Σ from
Normal-Inverse-Wishart (NIW) prior calculated in \code{niw.post()}.
}
\usage{
over_stat
}
\description{
Estimates of the percentage of overlap among example species used in
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\keyword{datasets}
