% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigma_est_wide}
\alias{sigma_est_wide}
\title{Estimates of sigma from Normal-Inverse-Wishart (NIW) prior}
\format{
data frame containing 8,0000 rows and 7 variables
\describe{
\item{metric}{name of the metric extracted from \code{niw.post()}}
\item{species}{species abbreviation}
\item{isotope}{column with isotope name}
\item{sample_number}{sample number from 1-1000}

\item{d15n}{estimate of sigma for d15n produced from \code{niw.post()}}
\item{d13c}{estimate of sigma for d13c produced from \code{niw.post()}}
}
}
\usage{
sigma_est_wide
}
\description{
Estimates of sigma from Normal-Inverse-Wishart (NIW) priors for example data
frame from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}
example
}
\keyword{datasets}
