% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-sigma_extract.R
\name{sigma_extract}
\alias{sigma_extract}
\title{Sigma extract}
\usage{
sigma_extract(data, isotope_a = NULL, isotope_b = NULL)
}
\arguments{
\item{data}{a list object created by \code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}}

\item{isotope_a}{character string to supply for the first
isotope used in \code{niw.post()}. Defaults to \code{"d15n"}.}

\item{isotope_b}{character string to supply for the second
isotope used in \code{niw.post()}. Defaults to \code{"d13c"}.}
}
\value{
Returns a \code{tibble} of extracted estimates of sigma that are created by
\code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
The tibble will contain five columns in the following order, \code{metric}, \code{id},
\code{sample_name}, \code{isotope}, \code{sample_number}, and the posterior sample for sigma.
(i.e., \code{post_sample})
}
\description{
Extract Bayesian estimates of sigma from the function \code{niw.post()} from
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
df_sigma <- sigma_extract(
data = niw_fish_post
)
# --- to use with `niche_ellipse()` we need to make into wide format ----
# we can do this by using tidyr's `pivot_wide()`. Uncomment below to run.
# df_sigma_wide <- df_sigma \%>\%
# pivot_wider(names_from = id,
#             values_from = post_sample)


# ---- To understand how niw_fish_post is being created ----
# ---- Uncomment the code below to create object from nicheROVER ----
# {
# library(dplyr)
# library(janitor)
# library(nicheROVER)
# library(nichetools)
# library(purrr)
# }
#
# gab fish dataframe, and remove sulfer for the example
# df <- fish \%>\%
#   janitor::clean_names()
#
# create number of samples used in nicheROVER
# nsample <- 1000
#
# split the data frame by species and select isotopes of interest
# df_split <- df \%>\%
# split(.$species) \%>\%
# map(~ select(., d15n, d13c))
#
# extract the names of each list to name each object in list
# df_names <- df \%>\%
# group_by(species) \%>\%
# group_keys() \%>\%
# ungroup() \%>\%
# mutate(
# id = 1:nrow(.)
# )
#
# name each object in list
# names(df_split) <- df_names$species
#
# create niw posterior samples
# niw_fish_post <- df_split \%>\%
# map(~niw.post(nsample = nsample, X = .))
#
#
# we can then use sigma_extract() to extract mu from niw_fish_post
# df_sigmaa <- sigma_extract(
# data = niw_fish_post
# )
#
# To prep this for use with


}
