% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-sigma_extract.R
\name{sigma_extract}
\alias{sigma_extract}
\title{Σ extract}
\usage{
sigma_extract(data, isotope_a = NULL, isotope_b = NULL, format = "wide")
}
\arguments{
\item{data}{a \code{list} created by the function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}}

\item{isotope_a}{\code{character} string to supply for the first
isotope used in \code{niw.post()}. Defaults to \code{"d15n"}.}

\item{isotope_b}{\code{character}string to supply for the second
isotope used in \code{niw.post()}. Defaults to \code{"d13c"}.}

\item{format}{format a \code{character} that decided whether the returned object is
in long or wide format. Default is \code{"wide"}, with the alternative supplied
being \code{"long"}.}
}
\value{
Returns a \code{tibble} of extracted estimates of Σ created by the
function \code{niw.post()} in the package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
The tibble will contain five columns in the following order, \code{metric}, \code{id},
\code{sample_name}, \code{isotope}, \code{sample_number}, and the posterior sample for Σ
(i.e., \code{post_sample}).
}
\description{
Extract Bayesian estimates of Σ from the function \code{niw.post()} in the
package
\href{https://cran.r-project.org/web/packages/nicheROVER/index.html}{{nicheROVER}}.
}
\examples{
sigma_extract(
data = niw_fish_post
)
# ---- below will be turned into vignette when finished ----
\dontrun{
# load packages
{
  library(dplyr)
  library(janitor)
  library(nicheROVER)
  library(nichetools)
  library(purrr)
  }

# ---- To understand how niw_fish_post is being created ----

# get fish data frame, and remove sulfur for the example
df <- fish \%>\%
    janitor::clean_names()

# create number of samples used in nicheROVER
nsample <- 1000

# split the data frame by species and select isotopes of interest
df_split <- df \%>\%
  split(.$species) \%>\%
  map(~ select(., d15n, d13c))

# extract the names of each list to name each object in list
df_names <- df \%>\%
   group_by(species) \%>\%
   group_keys() \%>\%
   ungroup() \%>\%
   mutate(
     id = 1:nrow(.)
     )

# name each object in list
names(df_split) <- df_names$species

# create niw posterior samples
niw_fish_post <- df_split \%>\%
  map(~niw.post(nsample = nsample, X = .))

# we can then use sigma_extract() to extract sigma from niw_fish_post
df_sigma <- sigma_extract(
  data = niw_fish_post
  )

# --- to use with `niche_ellipse()` we need to make into wide format ----
df_sigma_wide <- df_sigma |>
tidyr::pivot_wider(names_from = id,
            values_from = post_sample)
}
}
\seealso{
\code{\link[nicheROVER:niw.post]{nicheROVER::niw.post()}}
}
